+++ CHANGES.md
@@ -1,6 +1,8 @@
 # About CrateDB changelog
 
 ## Unreleased
+- Outline: Improved `llms-txt`'s `get_doc_content` to fail on
+  resources with HTTP != 200
 
 ## v0.0.8 - 2025-07-28
 - Outline: Shrank llms-txt output to <200_000 input tokens